Site: Set correct `lang` attribute for `<html>` tag
