+++ site/src/app/[domain]/[language]/layout.tsx
@@ -1,6 +1,9 @@
+import { SitePreviewProvider } from "@comet/cms-site";
 import { IntlProvider } from "@src/util/IntlProvider";
 import { loadMessages } from "@src/util/loadMessages";
 import { getSiteConfigForDomain } from "@src/util/siteConfig";
+import { SiteConfigProvider } from "@src/util/SiteConfigProvider";
+import { draftMode } from "next/headers";
 import { PropsWithChildren } from "react";
 
 export default async function Page({ children, params: { domain, language } }: PropsWithChildren<{ params: { domain: string; language: string } }>) {
@@ -9,10 +12,28 @@ export default async function Page({ children, params: { domain, language } }: P
         language = "en";
     }
 
+    const isDraftModeEnabled = draftMode().isEnabled;
+
     const messages = await loadMessages(language);
     return (
-        <IntlProvider locale={language} messages={messages}>
-            {children}
-        </IntlProvider>
+        <html lang={language}>
+            <body>
+                {siteConfig.gtmId && (
+                    <noscript>
+                        <iframe
+                            src={`https://www.googletagmanager.com/ns.html?id=${siteConfig.gtmId}`}
+                            height="0"
+                            width="0"
+                            style={{ display: "none", visibility: "hidden" }}
+                        />
+                    </noscript>
+                )}
+                <IntlProvider locale={language} messages={messages}>
+                    <SiteConfigProvider siteConfig={siteConfig}>
+                        {isDraftModeEnabled ? <SitePreviewProvider>{children}</SitePreviewProvider> : children}
+                    </SiteConfigProvider>
+                </IntlProvider>
+            </body>
+        </html>
     );
 }
