+++ site/src/app/not-found.tsx
@@ -1,0 +1,10 @@
+// TODO https://github.com/vivid-planet/comet-starter/pull/285
+export default function NotFound() {
+    return (
+        <html lang="en">
+            <body>
+                <h1>Not found</h1>
+            </body>
+        </html>
+    );
+}
